Updates
- Fixed askEmmaEditAns when entity not saving

- Fixed prerequisite in module creation ONLY
  -> stored the prerequisites in a session rather than storing in another table in db
  -> manipulate the date displaying on the prerequisite table and the module table
  -> Added simple validation

Example:

Scenario: "Baking 103" course prerequisite is "Baking 102" and "Baking 102" prerequisite is "Baking 101"

If i add "Baking 103" into prereq, it will add in all 3 courses and wont show them in the course list.

If i add "Baking 102" into prereq, it will add in "Baking 101" and wont show these 2 in the course list. "Baking 103" still will be displayed in the course list.

If i add in "Baking 101", "Baking 102" and "Baking 103" will still remain in the course list whilst "Baking 101" will be added into the prereq list and removed from the course list

Assuming all 3 courses are added in the prereq list,

if i remove "Baking 102", it will auto remove "Baking 103" and keep "Baking 101"

if i remove "Baking 101", it will remove everything from the prereq and all will be displayed in the course list

- Did not add new table into the DB but i added the example modules as data. Rafid should be able to use these data to test.